Fixed types and problem-difficulty names

+++ packages/lib/storage/kod-storage.ts
@@ -1,11 +1,12 @@
 const set = (key: string, value: any) => {
-    
+    if(typeof localStorage == "undefined") return;
     if(typeof value === 'object') value = JSON.stringify(value);
 
     localStorage.setItem(key, value);
 }
 
 const get = (key: string) => {
+    if(typeof localStorage == "undefined") return null;
     return localStorage.getItem(key);
 }
 
@@ -15,10 +16,12 @@ const getObject = <T> (key: string): T | null => {
 }
 
 const remove = (key: string) => {
+    if(typeof localStorage == "undefined") return;
     localStorage.removeItem(key);
 }
 
 const clear = () => {
+    if(typeof localStorage == "undefined") return;
     localStorage.clear();
 }
